requirements.txt
scrapy
sqlalchemy
fastapi
uvicorn 
pydantic
psycopg2-binary
pandas 
cachetools
python-dotenv